third_party/WebKit/LayoutTests/editing/text-iterator/basic-iteration-shadowdom-expected.txt

Unit tests for WebCore text iterator with shadow tree support enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addShadowTreeWithDivElement(div); range.selectNodeContents(internals.shadowRoot(div)); internals.rangeAsText(range) is "b"
PASS appendBrElement(internals.shadowRoot(div).childNodes[0]); range.selectNodeContents(internals.shadowRoot(div)); internals.rangeAsText(range) is "b\n"
PASS successfullyParsed is true

TEST COMPLETE
